+++ code/graphics/vulkan/VulkanFrame.h
@@ -15,7 +15,9 @@ class VulkanFrame {
 		vk::DeviceSize uniformBufferSize,
 		vk::DeviceSize uniformAlignment,
 		vk::DeviceSize vertexBufferSize,
-		vk::DeviceSize vertexAlignment);
+		vk::DeviceSize vertexAlignment,
+		vk::DeviceSize stagingBufferSize,
+		vk::DeviceSize stagingAlignment);
 
 	void wait_for_gpu();
 	void reset();
@@ -23,6 +25,7 @@ class VulkanFrame {
 	vk::CommandBuffer commandBuffer() const { return m_commandBuffer; }
 	VulkanRingBuffer& uniformBuffer() { return m_uniformRing; }
 	VulkanRingBuffer& vertexBuffer() { return m_vertexRing; }
+	VulkanRingBuffer& stagingBuffer() { return m_stagingRing; }
 
 	vk::Semaphore imageAvailable() const { return m_imageAvailable.get(); }
 	vk::Semaphore renderFinished() const { return m_renderFinished.get(); }
@@ -45,6 +48,7 @@ class VulkanFrame {
 
 	VulkanRingBuffer m_uniformRing;
 	VulkanRingBuffer m_vertexRing;
+	VulkanRingBuffer m_stagingRing;
 };
 
 } // namespace vulkan